+++ realm/realm-library/src/androidTest/java/io/realm/RealmObjectTests.java
@@ -16,6 +16,8 @@
 
 package io.realm;
 
+import android.support.test.annotation.UiThreadTest;
+import android.support.test.rule.UiThreadTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.After;
@@ -29,6 +31,7 @@ import java.io.FileNotFoundException;
 import java.lang.ref.WeakReference;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
@@ -75,6 +78,8 @@ public class RealmObjectTests {
     public final ExpectedException thrown = ExpectedException.none();
     @Rule
     public final RunInLooperThread looperThread = new RunInLooperThread();
+    @Rule
+    public final UiThreadTestRule uiThreadTestRule = new UiThreadTestRule();
 
     private Realm realm;
     private RealmConfiguration realmConfig;
@@ -1675,14 +1680,15 @@ public class RealmObjectTests {
     @Test
     @RunTestInLooperThread
     public void addChangeListener_returnedObjectOfCopyToRealmOrUpdate() {
-        looperThread.realm.beginTransaction();
-        looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
+        Realm realm = looperThread.realm;
+        realm.beginTransaction();
+        realm.createObject(AllTypesPrimaryKey.class, 1);
 
         AllTypesPrimaryKey allTypesPrimaryKey = new AllTypesPrimaryKey();
         allTypesPrimaryKey.setColumnLong(1);
         allTypesPrimaryKey.setColumnFloat(42f);
-        allTypesPrimaryKey = looperThread.realm.copyToRealmOrUpdate(allTypesPrimaryKey);
-        looperThread.realm.commitTransaction();
+        allTypesPrimaryKey = realm.copyToRealmOrUpdate(allTypesPrimaryKey);
+        realm.commitTransaction();
 
         allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
             @Override
@@ -1695,81 +1701,87 @@ public class RealmObjectTests {
 
     // The object should be added to HandlerController.realmObjects only when the first time addListener called.
     @Test
-    @RunTestInLooperThread
+    @UiThreadTest
     public void addChangeListener_shouldAddTheObjectToHandlerRealmObjects() {
-        looperThread.realm.beginTransaction();
-        AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
-        looperThread.realm.commitTransaction();
-        assertTrue(looperThread.realm.handlerController.realmObjects.isEmpty());
+        realm.beginTransaction();
+        AllTypesPrimaryKey allTypesPrimaryKey = realm.createObject(AllTypesPrimaryKey.class, 1);
+        realm.commitTransaction();
+        final Map<WeakReference<RealmObjectProxy>, RealmQuery<? extends RealmModel>> realmObjects =
+                realm.handlerController.realmObjects;
 
-        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+        assertTrue(realmObjects.isEmpty());
+
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
             @Override
-            public void onChange(RealmModel element) {
+            public void onChange(AllTypesPrimaryKey element) {
             }
         });
 
-        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
-        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+        assertEquals(1, realmObjects.size());
+        for (WeakReference<RealmObjectProxy> ref : realmObjects.keySet()) {
             assertTrue(ref.get() == allTypesPrimaryKey);
         }
-
-        looperThread.testComplete();
     }
 
     // The object should be added to HandlerController.realmObjects only once.
     @Test
-    @RunTestInLooperThread
+    @UiThreadTest
     public void addChangeListener_shouldNotAddDupEntriesToHandlerRealmObjects() {
-        looperThread.realm.beginTransaction();
-        AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
-        looperThread.realm.commitTransaction();
-        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+        realm.beginTransaction();
+        AllTypesPrimaryKey allTypesPrimaryKey = realm.createObject(AllTypesPrimaryKey.class, 1);
+        realm.commitTransaction();
+        final Map<WeakReference<RealmObjectProxy>, RealmQuery<? extends RealmModel>> realmObjects =
+                realm.handlerController.realmObjects;
+
+        for (WeakReference<RealmObjectProxy> ref : realmObjects.keySet()) {
             assertFalse(ref.get() == allTypesPrimaryKey);
         }
 
         // Add different listeners twice
-        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
             @Override
-            public void onChange(RealmModel element) {
+            public void onChange(AllTypesPrimaryKey element) {
             }
         });
-        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
             @Override
-            public void onChange(RealmModel element) {
+            public void onChange(AllTypesPrimaryKey element) {
             }
         });
 
-        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
-        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+        assertEquals(1, realmObjects.size());
+        for (WeakReference<RealmObjectProxy> ref : realmObjects.keySet()) {
             assertTrue(ref.get() == allTypesPrimaryKey);
         }
-        looperThread.testComplete();
     }
 
     // The object should not be added to HandlerController again after the async query loaded.
     @Test
     @RunTestInLooperThread
     public void addChangeListener_checkHandlerRealmObjectsWhenCallingOnAsyncObject() {
-        looperThread.realm.beginTransaction();
-        looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
-        looperThread.realm.commitTransaction();
+        Realm realm = looperThread.realm;
+        realm.beginTransaction();
+        realm.createObject(AllTypesPrimaryKey.class, 1);
+        realm.commitTransaction();
+        final Map<WeakReference<RealmObjectProxy>, RealmQuery<? extends RealmModel>> realmObjects =
+                realm.handlerController.realmObjects;
 
-        final AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.where(AllTypesPrimaryKey.class).findFirstAsync();
-        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+        final AllTypesPrimaryKey allTypesPrimaryKey = realm.where(AllTypesPrimaryKey.class).findFirstAsync();
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
             @Override
-            public void onChange(RealmModel element) {
-                allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+            public void onChange(AllTypesPrimaryKey element) {
+                allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
                     @Override
-                    public void onChange(RealmModel element) {
+                    public void onChange(AllTypesPrimaryKey element) {
 
                     }
                 });
-                assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
+                assertEquals(1, realmObjects.size());
                 looperThread.testComplete();
             }
         });
-        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
-        for (RealmQuery<? extends RealmModel> query : looperThread.realm.handlerController.realmObjects.values()) {
+        assertEquals(1, realmObjects.size());
+        for (RealmQuery<? extends RealmModel> query : realmObjects.values()) {
             assertNotNull(query);
         }
     }